added user schema & very mild refactoring
