+++ config.js
@@ -39,9 +39,8 @@ module.exports = {
             `${baseURL}/auth/google/redirect`
         ],
         scopes: [
-            'https://www.googleapis.com/auth/userinfo.email',
-            'https://www.googleapis.com/auth/userinfo.profile',
-            //'https://www.googleapis.com/auth/youtube.readonly'
+            //'profile', 
+            'email',
         ]
     },
 };
